+++ LeaveSystemMVC/Controllers/HumanResources/hrPendingApplicationController.cs
@@ -942,25 +942,6 @@ namespace LeaveSystemMVC.Controllers
             return leaveBalanceID;
         }
 
-        public void SendMail(string email, string message)
-        {
-            MailMessage mail = new MailMessage();
-            mail.From = new MailAddress("project_ict333@murdochdubai.ac.ae", "GIMEL LMS");
-            mail.To.Add(new MailAddress(email));
-            mail.Subject = "Leave Application Update";
-            mail.Body = message + Environment.NewLine;
-
-            SmtpClient client = new SmtpClient();
-            client.EnableSsl = true;
-            client.Credentials = new NetworkCredential("project_ict333@murdochdubai.ac.ae", "ict@333");
-            try
-            {
-                client.Send(mail);
-                System.Diagnostics.Debug.WriteLine("Mail Sent");
-            }
-            catch (Exception e){}
-        }
-
         private List<sLeaveModel> GetLeaveHistory(int empID)
         {
             var leaveHistory = new List<sLeaveModel>();
